Move dummy tile to common
