+++ common/dummy_core_magma.py
