+++ test_tile/test_tile_magma.py
@@ -1,4 +1,4 @@
-from test_tile.dummy_core_magma import DummyCore
+from common.dummy_core_magma import DummyCore
 import magma
 from bit_vector import BitVector
 from tile.tile_magma import Tile
@@ -9,5 +9,5 @@ def test_tile():
     core = DummyCore()
     tile = Tile(core)
     tile_circ = tile.circuit()
-    magma.compile(f"test_tile/build/{tile.name}", tile_circ,
+    magma.compile(f"{tile.name}", tile_circ,
                   output="coreir-verilog")
